course done (with a little extra)

+++ app/Http/Controllers/ListingController.php
@@ -63,6 +63,12 @@ class ListingController extends Controller
     // Update Form
 
     public function update(Request $request, Listings $listing){
+
+        //Make sure logged in user is owner
+        if($listing->users_id!=auth()->id()){
+            abort('403','Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title'=>'required',
             'company'=>['required'],
@@ -89,6 +95,10 @@ class ListingController extends Controller
     // Delete Form
 
     public function destroy(Listings $listing){
+         //Make sure logged in user is owner
+         if($listing->users_id!=auth()->id()){
+            abort('403','Unauthorized Action');
+        }
         $listing->delete();
 
         return redirect('/')->with('message','Listing Deleted Successfully');
